Reorganize repository for consistency with other AlbertaSat hardware repositories
