+++ equipment_handler/include/NMEAParser.h
