+++ equipment_handler/include/skytraq_binary.h
