+++ equipment_handler/source/NMEAParser.c
