+++ equipment_handler/source/skytraq_binary.c
