+++ hardware_interface/include/skytraq_gps.h
