+++ hardware_interface/source/housekeeping_charon.c
