+++ hardware_interface/test/charon_binary_test.h
